Update jupyterlab to version 4.3.3 and add libchebipy package
This commit updates the jupyterlab package from version 4.3.2 to 4.3.3, including the corresponding wheel and tar.gz files. Additionally, it introduces the libchebipy package (version 1.0.10) to the poetry.lock file, which provides a Python API for accessing the ChEBI database, along with its dependencies. The httpx dependency version has also been modified to allow for a broader range of compatible versions.